Fix Google Play SSL Warning
When uploading your app to Google Play it will be reject and you'll get this message:
"We rejected [app name], with package name [package name], for violating our Malicious Behavior or User Data policy. If you submitted an update, the previous version of your app is still available on Google Play.
This app uses software that contains security vulnerabilities for users or allows the collection of user data without proper disclosure."
Now it's fixed!